Add concept definition to UNIT_ADD_DIMENSION_TRAIT
Not sure whether the concept definition would be better placed in its
own macro, but this is an easy way to try out concept definitions
without having to change too much code.

The concepts naming convention from the C++ standard library (i.e.,
snake_case) is used here [0]. This is partially to stay consistent with
the standard library's naming conventions and partially for the sake of
convenience, as UNIT_ADD_DIMENSION_TRAIT is called with snake_case
arguments.

snake_case names, however, result in name clashes with
units::<dimension> namespaces, as well as with the units::dimensionless
type. This can be worked around using a concepts namespace. PascalCase
concept names is another solution, though that would require quite a few
more code changes to manually specify the concept names instead of
reusing the unitdimension parameter.

[0]: http://open-std.org/JTC1/SC22/WG21/docs/papers/2019/p1851r0.html